@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: taikai/v1/enums.proto
 
@@ -100,7 +100,7 @@ func file_taikai_v1_enums_proto_rawDescGZIP() []byte {
 }
 
 var file_taikai_v1_enums_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_taikai_v1_enums_proto_goTypes = []interface{}{
+var file_taikai_v1_enums_proto_goTypes = []any{
 	(HelloType)(0), // 0: taikai.v1.HelloType
 }
 var file_taikai_v1_enums_proto_depIdxs = []int32{
