@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: taikai/v1/requests.proto
 
@@ -418,8 +418,11 @@ type ListOrgRequest struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Name  *string `protobuf:"bytes,1,opt,name=name,proto3,oneof" json:"name,omitempty"`
-	OrgId *string `protobuf:"bytes,2,opt,name=orgId,proto3,oneof" json:"orgId,omitempty"`
+	// Types that are assignable to Filter:
+	//
+	//	*ListOrgRequest_Id
+	//	*ListOrgRequest_Name
+	Filter isListOrgRequest_Filter `protobuf_oneof:"filter"`
 }
 
 func (x *ListOrgRequest) Reset() {
@@ -454,19 +457,42 @@ func (*ListOrgRequest) Descriptor() ([]byte, []int) {
 	return file_taikai_v1_requests_proto_rawDescGZIP(), []int{8}
 }
 
-func (x *ListOrgRequest) GetName() string {
-	if x != nil && x.Name != nil {
-		return *x.Name
+func (m *ListOrgRequest) GetFilter() isListOrgRequest_Filter {
+	if m != nil {
+		return m.Filter
+	}
+	return nil
+}
+
+func (x *ListOrgRequest) GetId() string {
+	if x, ok := x.GetFilter().(*ListOrgRequest_Id); ok {
+		return x.Id
 	}
 	return ""
 }
 
-func (x *ListOrgRequest) GetOrgId() string {
-	if x != nil && x.OrgId != nil {
-		return *x.OrgId
+func (x *ListOrgRequest) GetName() string {
+	if x, ok := x.GetFilter().(*ListOrgRequest_Name); ok {
+		return x.Name
 	}
 	return ""
 }
+
+type isListOrgRequest_Filter interface {
+	isListOrgRequest_Filter()
+}
+
+type ListOrgRequest_Id struct {
+	Id string `protobuf:"bytes,1,opt,name=id,proto3,oneof"`
+}
+
+type ListOrgRequest_Name struct {
+	Name string `protobuf:"bytes,2,opt,name=name,proto3,oneof"`
+}
+
+func (*ListOrgRequest_Id) isListOrgRequest_Filter() {}
+
+func (*ListOrgRequest_Name) isListOrgRequest_Filter() {}
 
 type ListOrgResponse struct {
 	state         protoimpl.MessageState
@@ -614,8 +640,8 @@ type UpsertOrgUserRequest struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	OrgId *string `protobuf:"bytes,1,opt,name=orgId,proto3,oneof" json:"orgId,omitempty"`
-	User  *User   `protobuf:"bytes,2,opt,name=user,proto3,oneof" json:"user,omitempty"`
+	OrgId string `protobuf:"bytes,1,opt,name=orgId,proto3" json:"orgId,omitempty"`
+	User  *User  `protobuf:"bytes,2,opt,name=user,proto3" json:"user,omitempty"`
 }
 
 func (x *UpsertOrgUserRequest) Reset() {
@@ -651,8 +677,8 @@ func (*UpsertOrgUserRequest) Descriptor() ([]byte, []int) {
 }
 
 func (x *UpsertOrgUserRequest) GetOrgId() string {
-	if x != nil && x.OrgId != nil {
-		return *x.OrgId
+	if x != nil {
+		return x.OrgId
 	}
 	return ""
 }
@@ -669,7 +695,7 @@ type DeleteOrgRequest struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Id *string `protobuf:"bytes,1,opt,name=id,proto3,oneof" json:"id,omitempty"`
+	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 }
 
 func (x *DeleteOrgRequest) Reset() {
@@ -705,8 +731,8 @@ func (*DeleteOrgRequest) Descriptor() ([]byte, []int) {
 }
 
 func (x *DeleteOrgRequest) GetId() string {
-	if x != nil && x.Id != nil {
-		return *x.Id
+	if x != nil {
+		return x.Id
 	}
 	return ""
 }
@@ -744,44 +770,40 @@ var file_taikai_v1_requests_proto_rawDesc = []byte{
 	0x10, 0x55, 0x70, 0x73, 0x65, 0x72, 0x74, 0x4f, 0x72, 0x67, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
 	0x74, 0x12, 0x20, 0x0a, 0x03, 0x6f, 0x72, 0x67, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0e,
 	0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4f, 0x72, 0x67, 0x52, 0x03,
-	0x6f, 0x72, 0x67, 0x22, 0x57, 0x0a, 0x0e, 0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x17, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x88, 0x01, 0x01, 0x12, 0x19,
-	0x0a, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x01, 0x52,
-	0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x88, 0x01, 0x01, 0x42, 0x07, 0x0a, 0x05, 0x5f, 0x6e, 0x61,
-	0x6d, 0x65, 0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x22, 0x35, 0x0a, 0x0f,
-	0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
-	0x22, 0x0a, 0x04, 0x6f, 0x72, 0x67, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0e, 0x2e,
-	0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4f, 0x72, 0x67, 0x52, 0x04, 0x6f,
-	0x72, 0x67, 0x73, 0x22, 0x3a, 0x0a, 0x13, 0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73,
-	0x65, 0x72, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x19, 0x0a, 0x05, 0x6f, 0x72,
-	0x67, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x6f, 0x72, 0x67,
-	0x49, 0x64, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x22,
-	0x3d, 0x0a, 0x14, 0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73, 0x65, 0x72, 0x73, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x25, 0x0a, 0x05, 0x75, 0x73, 0x65, 0x72, 0x73,
-	0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0f, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e,
-	0x76, 0x31, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x52, 0x05, 0x75, 0x73, 0x65, 0x72, 0x73, 0x22, 0x6e,
-	0x0a, 0x14, 0x55, 0x70, 0x73, 0x65, 0x72, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73, 0x65, 0x72, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x19, 0x0a, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x88, 0x01,
-	0x01, 0x12, 0x28, 0x0a, 0x04, 0x75, 0x73, 0x65, 0x72, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x0f, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x55, 0x73, 0x65, 0x72,
-	0x48, 0x01, 0x52, 0x04, 0x75, 0x73, 0x65, 0x72, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f,
-	0x6f, 0x72, 0x67, 0x49, 0x64, 0x42, 0x07, 0x0a, 0x05, 0x5f, 0x75, 0x73, 0x65, 0x72, 0x22, 0x2e,
-	0x0a, 0x10, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x4f, 0x72, 0x67, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x12, 0x13, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00,
-	0x52, 0x02, 0x69, 0x64, 0x88, 0x01, 0x01, 0x42, 0x05, 0x0a, 0x03, 0x5f, 0x69, 0x64, 0x42, 0xa1,
-	0x01, 0x0a, 0x0d, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76, 0x31,
-	0x42, 0x0d, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x73, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50,
-	0x01, 0x5a, 0x3c, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x66, 0x6f,
-	0x72, 0x67, 0x65, 0x75, 0x74, 0x61, 0x68, 0x2f, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x67, 0x6f, 0x2f, 0x74, 0x61, 0x69,
-	0x6b, 0x61, 0x69, 0x2f, 0x76, 0x31, 0x3b, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x76, 0x31, 0xa2,
-	0x02, 0x03, 0x54, 0x58, 0x58, 0xaa, 0x02, 0x09, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x56,
-	0x31, 0xca, 0x02, 0x09, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x5c, 0x56, 0x31, 0xe2, 0x02, 0x15,
-	0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d, 0x65, 0x74,
-	0x61, 0x64, 0x61, 0x74, 0x61, 0xea, 0x02, 0x0a, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x3a, 0x3a,
-	0x56, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x6f, 0x72, 0x67, 0x22, 0x42, 0x0a, 0x0e, 0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x10, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x09, 0x48, 0x00, 0x52, 0x02, 0x69, 0x64, 0x12, 0x14, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x42, 0x08, 0x0a,
+	0x06, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x22, 0x35, 0x0a, 0x0f, 0x4c, 0x69, 0x73, 0x74, 0x4f,
+	0x72, 0x67, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x22, 0x0a, 0x04, 0x6f, 0x72,
+	0x67, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0e, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61,
+	0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4f, 0x72, 0x67, 0x52, 0x04, 0x6f, 0x72, 0x67, 0x73, 0x22, 0x3a,
+	0x0a, 0x13, 0x4c, 0x69, 0x73, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73, 0x65, 0x72, 0x73, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x19, 0x0a, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x88, 0x01, 0x01,
+	0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x22, 0x3d, 0x0a, 0x14, 0x4c, 0x69,
+	0x73, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73, 0x65, 0x72, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x73, 0x65, 0x12, 0x25, 0x0a, 0x05, 0x75, 0x73, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28,
+	0x0b, 0x32, 0x0f, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x55, 0x73,
+	0x65, 0x72, 0x52, 0x05, 0x75, 0x73, 0x65, 0x72, 0x73, 0x22, 0x51, 0x0a, 0x14, 0x55, 0x70, 0x73,
+	0x65, 0x72, 0x74, 0x4f, 0x72, 0x67, 0x55, 0x73, 0x65, 0x72, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x12, 0x14, 0x0a, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x05, 0x6f, 0x72, 0x67, 0x49, 0x64, 0x12, 0x23, 0x0a, 0x04, 0x75, 0x73, 0x65, 0x72, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0f, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e, 0x76,
+	0x31, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x52, 0x04, 0x75, 0x73, 0x65, 0x72, 0x22, 0x22, 0x0a, 0x10,
+	0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x4f, 0x72, 0x67, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64,
+	0x42, 0xa1, 0x01, 0x0a, 0x0d, 0x63, 0x6f, 0x6d, 0x2e, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x2e,
+	0x76, 0x31, 0x42, 0x0d, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x73, 0x50, 0x72, 0x6f, 0x74,
+	0x6f, 0x50, 0x01, 0x5a, 0x3c, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f,
+	0x66, 0x6f, 0x72, 0x67, 0x65, 0x75, 0x74, 0x61, 0x68, 0x2f, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69,
+	0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x67, 0x6f, 0x2f, 0x74,
+	0x61, 0x69, 0x6b, 0x61, 0x69, 0x2f, 0x76, 0x31, 0x3b, 0x74, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x76,
+	0x31, 0xa2, 0x02, 0x03, 0x54, 0x58, 0x58, 0xaa, 0x02, 0x09, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69,
+	0x2e, 0x56, 0x31, 0xca, 0x02, 0x09, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x5c, 0x56, 0x31, 0xe2,
+	0x02, 0x15, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69, 0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d,
+	0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0xea, 0x02, 0x0a, 0x54, 0x61, 0x69, 0x6b, 0x61, 0x69,
+	0x3a, 0x3a, 0x56, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -797,7 +819,7 @@ func file_taikai_v1_requests_proto_rawDescGZIP() []byte {
 }
 
 var file_taikai_v1_requests_proto_msgTypes = make([]protoimpl.MessageInfo, 14)
-var file_taikai_v1_requests_proto_goTypes = []interface{}{
+var file_taikai_v1_requests_proto_goTypes = []any{
 	(*Empty)(nil),                // 0: taikai.v1.Empty
 	(*ListRequest)(nil),          // 1: taikai.v1.ListRequest
 	(*DeleteRequest)(nil),        // 2: taikai.v1.DeleteRequest
@@ -838,7 +860,7 @@ func file_taikai_v1_requests_proto_init() {
 	file_taikai_v1_types_proto_init()
 	file_taikai_v1_enums_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_taikai_v1_requests_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Empty); i {
 			case 0:
 				return &v.state
@@ -850,7 +872,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ListRequest); i {
 			case 0:
 				return &v.state
@@ -862,7 +884,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteRequest); i {
 			case 0:
 				return &v.state
@@ -874,7 +896,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteResponse); i {
 			case 0:
 				return &v.state
@@ -886,7 +908,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetRequest); i {
 			case 0:
 				return &v.state
@@ -898,7 +920,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*Hellos); i {
 			case 0:
 				return &v.state
@@ -910,7 +932,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UpsertHellosRequest); i {
 			case 0:
 				return &v.state
@@ -922,7 +944,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*UpsertOrgRequest); i {
 			case 0:
 				return &v.state
@@ -934,7 +956,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*ListOrgRequest); i {
 			case 0:
 				return &v.state
@@ -946,7 +968,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*ListOrgResponse); i {
 			case 0:
 				return &v.state
@@ -958,7 +980,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*ListOrgUsersRequest); i {
 			case 0:
 				return &v.state
@@ -970,7 +992,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*ListOrgUsersResponse); i {
 			case 0:
 				return &v.state
@@ -982,7 +1004,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*UpsertOrgUserRequest); i {
 			case 0:
 				return &v.state
@@ -994,7 +1016,7 @@ func file_taikai_v1_requests_proto_init() {
 				return nil
 			}
 		}
-		file_taikai_v1_requests_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_taikai_v1_requests_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteOrgRequest); i {
 			case 0:
 				return &v.state
@@ -1007,10 +1029,11 @@ func file_taikai_v1_requests_proto_init() {
 			}
 		}
 	}
-	file_taikai_v1_requests_proto_msgTypes[8].OneofWrappers = []interface{}{}
-	file_taikai_v1_requests_proto_msgTypes[10].OneofWrappers = []interface{}{}
-	file_taikai_v1_requests_proto_msgTypes[12].OneofWrappers = []interface{}{}
-	file_taikai_v1_requests_proto_msgTypes[13].OneofWrappers = []interface{}{}
+	file_taikai_v1_requests_proto_msgTypes[8].OneofWrappers = []any{
+		(*ListOrgRequest_Id)(nil),
+		(*ListOrgRequest_Name)(nil),
+	}
+	file_taikai_v1_requests_proto_msgTypes[10].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
