@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.4.0
 // - protoc             (unknown)
 // source: taikai/v1/api.proto
 
@@ -11,12 +11,13 @@ import (
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.62.0 or later.
+const _ = grpc.SupportPackageIsVersion8
 
 const (
 	Api_UpsertHellos_FullMethodName  = "/taikai.v1.Api/UpsertHellos"
@@ -41,12 +42,12 @@ type ApiClient interface {
 	ListHellos(ctx context.Context, in *ListRequest, opts ...grpc.CallOption) (*Hellos, error)
 	GetHellos(ctx context.Context, in *GetRequest, opts ...grpc.CallOption) (*Hellos, error)
 	// Health check
-	Healthy(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error)
+	Healthy(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*emptypb.Empty, error)
 	// Readiness check
-	Ready(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error)
-	UpsertOrg(ctx context.Context, in *UpsertOrgRequest, opts ...grpc.CallOption) (*Empty, error)
+	Ready(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*emptypb.Empty, error)
+	UpsertOrg(ctx context.Context, in *UpsertOrgRequest, opts ...grpc.CallOption) (*emptypb.Empty, error)
 	ListOrg(ctx context.Context, in *ListOrgRequest, opts ...grpc.CallOption) (*ListOrgResponse, error)
-	UpsertOrgUser(ctx context.Context, in *UpsertOrgUserRequest, opts ...grpc.CallOption) (*Empty, error)
+	UpsertOrgUser(ctx context.Context, in *UpsertOrgUserRequest, opts ...grpc.CallOption) (*emptypb.Empty, error)
 	ListOrgUsers(ctx context.Context, in *ListOrgUsersRequest, opts ...grpc.CallOption) (*ListOrgUsersResponse, error)
 	DeleteOrg(ctx context.Context, in *DeleteRequest, opts ...grpc.CallOption) (*DeleteResponse, error)
 }
@@ -60,8 +61,9 @@ func NewApiClient(cc grpc.ClientConnInterface) ApiClient {
 }
 
 func (c *apiClient) UpsertHellos(ctx context.Context, in *UpsertHellosRequest, opts ...grpc.CallOption) (*Hellos, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hellos)
-	err := c.cc.Invoke(ctx, Api_UpsertHellos_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_UpsertHellos_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +71,9 @@ func (c *apiClient) UpsertHellos(ctx context.Context, in *UpsertHellosRequest, o
 }
 
 func (c *apiClient) DeleteHellos(ctx context.Context, in *DeleteRequest, opts ...grpc.CallOption) (*DeleteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteResponse)
-	err := c.cc.Invoke(ctx, Api_DeleteHellos_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_DeleteHellos_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -78,8 +81,9 @@ func (c *apiClient) DeleteHellos(ctx context.Context, in *DeleteRequest, opts ..
 }
 
 func (c *apiClient) ListHellos(ctx context.Context, in *ListRequest, opts ...grpc.CallOption) (*Hellos, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hellos)
-	err := c.cc.Invoke(ctx, Api_ListHellos_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_ListHellos_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -87,35 +91,39 @@ func (c *apiClient) ListHellos(ctx context.Context, in *ListRequest, opts ...grp
 }
 
 func (c *apiClient) GetHellos(ctx context.Context, in *GetRequest, opts ...grpc.CallOption) (*Hellos, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Hellos)
-	err := c.cc.Invoke(ctx, Api_GetHellos_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_GetHellos_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *apiClient) Healthy(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
-	out := new(Empty)
-	err := c.cc.Invoke(ctx, Api_Healthy_FullMethodName, in, out, opts...)
+func (c *apiClient) Healthy(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*emptypb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(emptypb.Empty)
+	err := c.cc.Invoke(ctx, Api_Healthy_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *apiClient) Ready(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
-	out := new(Empty)
-	err := c.cc.Invoke(ctx, Api_Ready_FullMethodName, in, out, opts...)
+func (c *apiClient) Ready(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*emptypb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(emptypb.Empty)
+	err := c.cc.Invoke(ctx, Api_Ready_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *apiClient) UpsertOrg(ctx context.Context, in *UpsertOrgRequest, opts ...grpc.CallOption) (*Empty, error) {
-	out := new(Empty)
-	err := c.cc.Invoke(ctx, Api_UpsertOrg_FullMethodName, in, out, opts...)
+func (c *apiClient) UpsertOrg(ctx context.Context, in *UpsertOrgRequest, opts ...grpc.CallOption) (*emptypb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(emptypb.Empty)
+	err := c.cc.Invoke(ctx, Api_UpsertOrg_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -123,17 +131,19 @@ func (c *apiClient) UpsertOrg(ctx context.Context, in *UpsertOrgRequest, opts ..
 }
 
 func (c *apiClient) ListOrg(ctx context.Context, in *ListOrgRequest, opts ...grpc.CallOption) (*ListOrgResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListOrgResponse)
-	err := c.cc.Invoke(ctx, Api_ListOrg_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_ListOrg_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *apiClient) UpsertOrgUser(ctx context.Context, in *UpsertOrgUserRequest, opts ...grpc.CallOption) (*Empty, error) {
-	out := new(Empty)
-	err := c.cc.Invoke(ctx, Api_UpsertOrgUser_FullMethodName, in, out, opts...)
+func (c *apiClient) UpsertOrgUser(ctx context.Context, in *UpsertOrgUserRequest, opts ...grpc.CallOption) (*emptypb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(emptypb.Empty)
+	err := c.cc.Invoke(ctx, Api_UpsertOrgUser_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -141,8 +151,9 @@ func (c *apiClient) UpsertOrgUser(ctx context.Context, in *UpsertOrgUserRequest,
 }
 
 func (c *apiClient) ListOrgUsers(ctx context.Context, in *ListOrgUsersRequest, opts ...grpc.CallOption) (*ListOrgUsersResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListOrgUsersResponse)
-	err := c.cc.Invoke(ctx, Api_ListOrgUsers_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_ListOrgUsers_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -150,8 +161,9 @@ func (c *apiClient) ListOrgUsers(ctx context.Context, in *ListOrgUsersRequest, o
 }
 
 func (c *apiClient) DeleteOrg(ctx context.Context, in *DeleteRequest, opts ...grpc.CallOption) (*DeleteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteResponse)
-	err := c.cc.Invoke(ctx, Api_DeleteOrg_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Api_DeleteOrg_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -167,12 +179,12 @@ type ApiServer interface {
 	ListHellos(context.Context, *ListRequest) (*Hellos, error)
 	GetHellos(context.Context, *GetRequest) (*Hellos, error)
 	// Health check
-	Healthy(context.Context, *Empty) (*Empty, error)
+	Healthy(context.Context, *emptypb.Empty) (*emptypb.Empty, error)
 	// Readiness check
-	Ready(context.Context, *Empty) (*Empty, error)
-	UpsertOrg(context.Context, *UpsertOrgRequest) (*Empty, error)
+	Ready(context.Context, *emptypb.Empty) (*emptypb.Empty, error)
+	UpsertOrg(context.Context, *UpsertOrgRequest) (*emptypb.Empty, error)
 	ListOrg(context.Context, *ListOrgRequest) (*ListOrgResponse, error)
-	UpsertOrgUser(context.Context, *UpsertOrgUserRequest) (*Empty, error)
+	UpsertOrgUser(context.Context, *UpsertOrgUserRequest) (*emptypb.Empty, error)
 	ListOrgUsers(context.Context, *ListOrgUsersRequest) (*ListOrgUsersResponse, error)
 	DeleteOrg(context.Context, *DeleteRequest) (*DeleteResponse, error)
 }
@@ -193,19 +205,19 @@ func (UnimplementedApiServer) ListHellos(context.Context, *ListRequest) (*Hellos
 func (UnimplementedApiServer) GetHellos(context.Context, *GetRequest) (*Hellos, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetHellos not implemented")
 }
-func (UnimplementedApiServer) Healthy(context.Context, *Empty) (*Empty, error) {
+func (UnimplementedApiServer) Healthy(context.Context, *emptypb.Empty) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Healthy not implemented")
 }
-func (UnimplementedApiServer) Ready(context.Context, *Empty) (*Empty, error) {
+func (UnimplementedApiServer) Ready(context.Context, *emptypb.Empty) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Ready not implemented")
 }
-func (UnimplementedApiServer) UpsertOrg(context.Context, *UpsertOrgRequest) (*Empty, error) {
+func (UnimplementedApiServer) UpsertOrg(context.Context, *UpsertOrgRequest) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpsertOrg not implemented")
 }
 func (UnimplementedApiServer) ListOrg(context.Context, *ListOrgRequest) (*ListOrgResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListOrg not implemented")
 }
-func (UnimplementedApiServer) UpsertOrgUser(context.Context, *UpsertOrgUserRequest) (*Empty, error) {
+func (UnimplementedApiServer) UpsertOrgUser(context.Context, *UpsertOrgUserRequest) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpsertOrgUser not implemented")
 }
 func (UnimplementedApiServer) ListOrgUsers(context.Context, *ListOrgUsersRequest) (*ListOrgUsersResponse, error) {
@@ -299,7 +311,7 @@ func _Api_GetHellos_Handler(srv interface{}, ctx context.Context, dec func(inter
 }
 
 func _Api_Healthy_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(Empty)
+	in := new(emptypb.Empty)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -311,13 +323,13 @@ func _Api_Healthy_Handler(srv interface{}, ctx context.Context, dec func(interfa
 		FullMethod: Api_Healthy_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ApiServer).Healthy(ctx, req.(*Empty))
+		return srv.(ApiServer).Healthy(ctx, req.(*emptypb.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _Api_Ready_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(Empty)
+	in := new(emptypb.Empty)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -329,7 +341,7 @@ func _Api_Ready_Handler(srv interface{}, ctx context.Context, dec func(interface
 		FullMethod: Api_Ready_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ApiServer).Ready(ctx, req.(*Empty))
+		return srv.(ApiServer).Ready(ctx, req.(*emptypb.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
